+++ app/models/dataset/credential/CredentialDAO.scala
@@ -97,7 +97,8 @@ class CredentialDAO @Inject()(sqlClient: SqlClient)(implicit ec: ExecutionContex
         case CredentialType.HttpBasicAuth        => parseAsHttpBasicAuthCredential(r)
         case CredentialType.S3AccessKey          => parseAsS3AccessKeyCredential(r)
         case CredentialType.GoogleServiceAccount => parseAsGoogleServiceAccountCredential(r)
-        case _                                   => Fox.failure(s"Unknown credential type: ${r.`type`}")
+        // Keep in sync with config reader methods in CredentialConfigReader
+        case _ => Fox.failure(s"Unknown credential type: ${r.`type`}")
       }
     } yield parsed
 }
