+++ conf/application.conf
@@ -228,20 +228,12 @@ datastore {
   }
   agglomerateSkeleton.maxEdges = 1000000
   dataVaults {
-    credentials = [
-      {
-        type = "S3AccessKey"
-        name = "s3://myBucketb3c357e8-cab8-41fb-9ccd-f390197a6793"
-        identifier = "someKeyId"
-        secret = "someSecret"
-      },
-      {
-        type = "S3AccessKey"
-        name = "s3://anotherBucketb3c357e8-cab8-41fb-9ccd-f390197a6793"
-        identifier = "anotherKeyId"
-        secret = "anotherSecret"
-      }
-    ]
+    # Global data vault access credentials. Not available during explore, only during data loading.
+    # Example: {type = "S3AccessKey", name = "s3://example/uri/prefix", identifier = "someKeyId", secret = "someSecret"}
+    # type can be one of HttpBasicAuth, S3AccessKey, GoogleServiceAccount
+    # For GoogleServiceAccount the secret is a json string.
+    # The credentials are selected by uri prefix, so different s3 uri styles may need duplicated credential entries.
+    credentials = []
   }
 }
