+++ webknossos-datastore/app/com/scalableminds/webknossos/datastore/storage/CredentialConfigReader.scala
@@ -1,0 +1,65 @@
+package com.scalableminds.webknossos.datastore.storage
+
+import com.scalableminds.util.tools.ConfigReader
+import com.typesafe.config.Config
+import net.liftweb.common.Box.tryo
+import play.api.Configuration
+import play.api.libs.json.Json
+
+class CredentialConfigReader(underlyingConfig: Config) extends ConfigReader {
+  override val raw: Configuration = Configuration(underlyingConfig)
+
+  def getCredential: Option[DataVaultCredential] =
+    for {
+      typeLiteral <- getOptional[String]("type")
+      typParsed <- CredentialType.fromString(typeLiteral)
+      credential <- typParsed match {
+        case CredentialType.S3AccessKey          => getAsS3
+        case CredentialType.HttpBasicAuth        => getAsHttpsBasicAuth
+        case CredentialType.GoogleServiceAccount => getAsGoogleServiceAccount
+        // Keep in sync with parse methods in CredentialDAO
+        case _ => None
+      }
+    } yield credential
+
+  private def getAsS3: Option[S3AccessKeyCredential] =
+    for {
+      name <- getOptional[String]("name")
+      keyId <- getOptional[String]("identifier")
+      key <- getOptional[String]("secret")
+    } yield
+      S3AccessKeyCredential(
+        name,
+        keyId,
+        key,
+        None,
+        None
+      )
+
+  private def getAsHttpsBasicAuth: Option[HttpBasicAuthCredential] =
+    for {
+      name <- getOptional[String]("name")
+      username <- getOptional[String]("identifier")
+      password <- getOptional[String]("secret")
+    } yield
+      HttpBasicAuthCredential(
+        name,
+        username,
+        password,
+        None,
+        None
+      )
+
+  private def getAsGoogleServiceAccount: Option[GoogleServiceAccountCredential] =
+    for {
+      name <- getOptional[String]("name")
+      secret <- getOptional[String]("secret")
+      secretJson <- tryo(Json.parse(secret)).toOption
+    } yield
+      GoogleServiceAccountCredential(
+        name,
+        secretJson,
+        None,
+        None
+      )
+}
